backend/requirements.txt
flask
flask-socketio
pandas
requests
python-dotenv
googlemaps